[SPARK-22672][TEST][SQL] Move OrcTest to `sql/core`
